requirements.txt
pyyaml
graphviz
pandas
matplotlib
numpy
scikit-learn
scipy
seaborn
sqlalchemy
pymysql
mysql-connector-python
